Refactored all reusable stuffs into lease.go
